imt: Add support for hybrid mode
In Hybrid mode, the number of contacts that can be reported in one report
is less than the maximum number of contacts that the device supports.
For example, a device that supports a maximum of 4 concurrent physical
contacts, can set up its top-level collection to deliver a maximum of
two contacts in one report. If four contact points are present, the device
can break these up into two serial reports that deliver two contacts each.

"When a device delivers data in this manner, the Contact Count usage value
in the first report should reflect the total number of contacts that are
being delivered in the hybrid reports. The other serial reports should have
a contact count of zero (0)."