[DRAFT] iichid: add support for gpio interrupt
At the moment there is no support for GPIO interrupts on x86, but there is
some work in progress.
This change shows how to obtain GPIO interrupt configuration from ACPI
and to set up scuh an interrupt in a consumer driver.
The GPIO side of things will probably have some changes as the work
progresses.  The ACPI related code is probably fine already.