[hkbd] Fix key repeats on multireport keyboards. Take 3.
Currently hkbd counts all key states to be "Up" at the start of
interrupt callback. That results in generation of "Key Up" event for
each key that has been downed before but is not listed in current
report while still is downed.

Fix that with clearing of temporary key data storage bits only for
keys contained in processed report.